pyproject.toml
[tool.poetry]
name = "demofastapi_poetry"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.22.3"
python-dotenv = "^0.20.0"
fastapi = "^0.76.0"
uvicorn = "^0.17.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
